+++ src/AutoNumeric.js
@@ -1,8 +1,8 @@
 /**
  *               AutoNumeric.js
  *
- * @version      4.5.8
- * @date         2019-10-08 UTC 00:34
+ * @version      4.5.9
+ * @date         2019-10-12 UTC 07:07
  *
  * @authors      Bob Knothe, Alexandre Bonneau
  * @contributors Sokolov Yura and others, cf. AUTHORS
@@ -909,7 +909,7 @@ export default class AutoNumeric {
      * @returns {string}
      */
     static version() {
-        return '4.5.8';
+        return '4.5.9';
     }
 
     /**
@@ -9008,7 +9008,12 @@ To solve that, you'd need to either set \`decimalPlacesRawValue\` to \`null\`, o
                 }
             }
 
-            const leftReg = new RegExp('^.*?' + leftAr.join('.*?'));
+            let leftReg;
+            if (this.settings.currencySymbolPlacement === AutoNumeric.options.currencySymbolPlacement.suffix) {
+                leftReg = new RegExp(`^.*?${leftAr.join('.*?')}`);
+            } else { // prefix is assumed
+                leftReg = new RegExp(`^.*${this.settings.currencySymbol}.*${leftAr.join('.*?')}`); // Fixes issue #647 when using a currency that has some characters in it that matches the value we just entered (ie. numbers in the currency)
+            }
 
             // Search cursor position in formatted value
             const newLeft = value.match(leftReg);